Add KSP-AVC version file
This commit adds a version file according to the KSP-AVC schema:
https://github.com/linuxgurugamer/KSPAddonVersionChecker/blob/master/KSP-AVC.schema.json

The file should be included in the LMPClient in future releases. The AppVeyor script has been adjusted accordingly.

It allows users to track available updates and compatibility of LMP if they have the "Addon Version Checker" mod installed.

It also allows this mod to be indexed on CKAN, because the NetKAN bot can parse the compatibility data from this file.